pkgs/rxvt_unicode_24bit: Remove, no longer required
rxvt-unicode 9.30 has a 24-bit colour implementation. It's dynamically
paletted, which is unfortunate (as it won't work correctly with e.g.
sixel images, if that ever becomes supported), but it works OK for my
current use-cases. The font-width and line-spacing patches are also no
longer required, as upstream has good-enough fixes for the issues now.